refactor: doc-code alignment — fix module descriptions, move file_hash, split large test files
DESIGN.md: fix module descriptions (D1-D6) — compliance pipeline, CheckType/Requirement,
commands.py orchestration hub, refactor/ submodule detail, fs.py with file_hash/safe_parse_py.
PATTERNS.md: fix AssistantBackend protocol (3→7 members), stale signatures (safe_parse_py,
create_agent_symlinks), Session Command Wrapper description, Shared Utility module boundaries.
Code: move file_hash from config.py to fs.py, update all import sites.
Tests: create 5 missing test files (adoption_templates, fs, checks/adoption, refactor/models,
refactor/testability). Split 4 large test files (promise, cli, compliance, checks) — all under 500 lines.
