+++ docs/PATTERNS.md
@@ -39,8 +39,15 @@ To support multiple AI assistants (Claude Code, opencode, Gemini CLI) with ident
 
 ```python
 class AssistantBackend(Protocol):
+    @property
+    def name(self) -> str: ...
+    @property
+    def cli_command(self) -> str: ...
+    @property
+    def install_hint(self) -> str: ...
     def build_command(self, skill_name: str, cwd: Path, model: str | None = None) -> list[str]: ...
     def sync_skills(self) -> None: ...
+    def env_overrides(self) -> dict[str, str]: ...
     def subagent_type_map(self) -> dict[str, str]: ...
 ```
 
@@ -48,12 +55,13 @@ class AssistantBackend(Protocol):
 Following the "Test Value Over Test Count" philosophy, the system uses AST heuristics to identify "testable logic" (e.g., non-trivial branching or calculations). This allows automated tools to detect missing tests only for modules that truly require them, avoiding redundant coverage of trivial pass-through code.
 
 ```python
-def _has_testable_logic(path: Path) -> bool: ...
+def _is_trivial_function(node: ast.FunctionDef | ast.AsyncFunctionDef) -> bool: ...
+def _is_testable_function(node: ast.FunctionDef | ast.AsyncFunctionDef) -> bool: ...
 def _is_testable_class(node: ast.ClassDef) -> bool: ...
 ```
 
 ### Session Command Wrapper Pattern
-To ensure consistent error handling across all six CLI session commands (`spec`, `design`, `patterns`, `execute`, `compliance`, `refactor`), the system uses a centralized wrapper to catch domain-specific errors and present them uniformly through the terminal UI.
+To ensure consistent error handling across all six CLI session commands (`spec`, `design`, `patterns`, `execute`, `compliance`, `refactor`), the system uses a centralized wrapper to catch domain-specific errors and present them via `typer.echo(str(exc), err=True)`.
 
 ```python
 def _run_session_command(cmd: Callable[..., int | None], agent: str | None, model: str | None, provider: str | None) -> None: ...
@@ -76,8 +84,15 @@ def init_project(cwd: Path | None = None) -> None: ...
 File operations that scan large directory trees or load non-essential configuration use a "silently degrade" pattern. By catching `OSError` and returning a safe default (like an empty list or string), the system maintains stability during batch operations.
 
 ```python
+# fs.py utilities
 def file_hash(path: Path) -> str | None: ...
-def collect_module_metrics(root: Path) -> list[ModuleMetrics]: ...
+def safe_parse_py(path: Path) -> tuple[ast.Module, str] | None: ...
+def atomic_write(target: Path, data: bytes) -> None: ...
+def create_agent_symlinks(root: Path, agents_path: Path) -> list[Path]: ...
+def xdg_config_home() -> Path: ...
+
+# checks/utils.py utilities
+def check_path_exists(path: Path, *, negative: bool = False) -> None: ...
 ```
 
 ### Terminal Failure Pattern
